Refactor into separate file, add cssProp option
